refactor(*:skip) Move docs from `doc` to `docs` (#4660)
