+++ .github/workflows/baselines.yml
@@ -36,7 +36,7 @@ jobs:
                 FILTER+=$(echo "$DIR: ${BASELINES_PATH}/**\n")
             done < <(find baselines -maxdepth 1 \
                 -name ".*" -prune -o \
-                -path "baselines/doc" -prune -o \
+                -path "baselines/docs" -prune -o \
                 -path "baselines/dev" -prune -o \
                 -path "baselines/baseline_template" -prune -o \
                 -path "baselines/flwr_baselines" -prune -o \
