+++ baselines/docs/source/.gitignore
