+++ datasets/docs/source/_static/readme/comparison_of_partitioning_schemes.png
