+++ datasets/docs/source/_templates/autosummary/base.rst
