+++ datasets/docs/source/_templates/autosummary/class.rst
