+++ datasets/docs/source/_templates/autosummary/module.rst
