+++ dev/update_version.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 
 REPLACE_CURR_VERSION = {
-    "doc/source/conf.py": [
+    "framework/docs/source/conf.py": [
         ".. |stable_flwr_version| replace:: {version}",
     ],
     "src/py/flwr/cli/new/templates/app/pyproject.*.toml.tpl": [
@@ -17,11 +17,11 @@ REPLACE_CURR_VERSION = {
 
 REPLACE_NEXT_VERSION = {
     "pyproject.toml": ['version = "{version}"'],
-    "doc/source/conf.py": [
+    "framework/docs/source/conf.py": [
         'release = "{version}"',
     ],
-    "examples/doc/source/conf.py": ['release = "{version}"'],
-    "baselines/doc/source/conf.py": ['release = "{version}"'],
+    "examples/docs/source/conf.py": ['release = "{version}"'],
+    "baselines/docs/source/conf.py": ['release = "{version}"'],
     "src/docker/complete/compose.yml": ["FLWR_VERSION:-{version}"],
     "src/docker/distributed/client/compose.yml": ["FLWR_VERSION:-{version}"],
     "src/docker/distributed/server/compose.yml": ["FLWR_VERSION:-{version}"],
