+++ examples/docs/Makefile
