+++ examples/docs/make.bat
