+++ examples/docs/source/_static/.gitignore
