+++ examples/docs/source/_static/favicon.ico
