+++ examples/docs/source/_static/flower-logo.png
